Move file formatter to separate module
